requirements.txt
prettytable
openpyxl
transformers
requests
dnspython
